+++ lib/db/migrations/0001_petite_titania.sql
@@ -1,0 +1,8 @@
+ALTER TABLE "users" ADD COLUMN "first_name" varchar(255);--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "last_name" varchar(255);--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "avatar_url" text;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "phone_number" varchar(20);--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "telegram_username" varchar(255);--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "organization_id" uuid;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "role_id" uuid;--> statement-breakpoint
+ALTER TABLE "users" ADD COLUMN "team_id" uuid;