+++ utils/supabase/server.ts
@@ -1,0 +1,41 @@
+import { createServerClient } from '@supabase/ssr'
+import { cookies } from 'next/headers'
+import { Database } from '@/types/supabase'
+
+export function createClient() {
+  const cookieStore = cookies()
+  
+  return createServerClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
+    {
+      cookies: {
+        async get(name) {
+          const cookieJar = await cookieStore
+          const cookie = cookieJar.get(name)
+          return cookie?.value
+        },
+        async set(name, value, options) {
+          try {
+            const cookieJar = await cookieStore
+            cookieJar.set(name, value, options)
+          } catch (error) {
+            // The `set` method was called from a Server Component.
+            // This can be ignored if you have middleware refreshing
+            // user sessions.
+          }
+        },
+        async remove(name, options) {
+          try {
+            const cookieJar = await cookieStore
+            cookieJar.set(name, '', { ...options, maxAge: 0 })
+          } catch (error) {
+            // The `delete` method was called from a Server Component.
+            // This can be ignored if you have middleware refreshing
+            // user sessions.
+          }
+        },
+      },
+    }
+  )
+}